Improve handling of merge conflicts when updating from upstream repos
Previously we would fail to update from a repo if it contained any merge
conflicts at all.  Now we only fail if there is merge conflict in the
tests themselves.

This current unblock the merging of reference-types, threads, and
multi-value.  The only repo with merge conflicts in the tests themselves
is sign-extension-ops.