Fix expected validation errors for `elem.drop`
The validation for `elem.drop` only checks that the referenced segment is valid within the context, it does not check the existence of tables. Therefore, when asserting that a module is invalid because of a bad segment index in `elem.drop`, we should expect an error message that references an unknown element segment, not a message that references an unknown table.