Add a CI workflow for auto-updating this repository
This commit adds `.github/workflows/autoupdate.yml` to this repository
which is scheduled to run once a week on Wednesday around noon PST
(assuming I got the time zones right). This workflow will run the
`./update-testsuite.sh` script and if it makes a commit it'll make a PR
to this repository.

The motivation for this is to ensure that this repository doesn't lag
behind too far upstream proposals and making it as easy as possible to
land updates (ideally "just hit the big green button").

This commit updates some words in the `README.md` about the update
process and additionally updates the script to not fail if no tests need
an update.

An example pull request [looks like this][pr] and an example workflow
which runs but didn't actually end up updating anything [looks like
this][noop]

[pr]: https://github.com/alexcrichton/testsuite/pull/1
[noop]: https://github.com/alexcrichton/testsuite/actions/runs/9522240608/job/26251380263