Rewrite the update script in Python
With various updates to this script such as having a different upstream
for `memory64` and otherwise including the `wasm-3.0` test suite this
script was getting relatively complicated which is typically not a great
situation for any bash script to be in. Instead rewrite this script in
Python to avoid many of bash's pitfalls and be a bit more
readable/reviewable/modifiable.