Remove some archived proposals from this repository
This commit removes the test suite mirror for the following proposals:

* `exception-handling`
* `extended-const`
* `function-references`
* `gc`
* `multi-memory`
* `tail-call`

These proposals have all been merged into the `wasm-3.0` branch of the
specification and their corresponding proposal repositories have all
been archived. My assumption is that this means that the source of truth
for these specifications now lives in the `wasm-3.0` branch and thus
don't need to be doubly-included as those tests are already included.